Move routes to components
