+++ client/src/components/App.tsx
@@ -4,7 +4,7 @@ import {BrowserRouter} from "react-router-dom";
 
 import Header from "./Header";
 import Footer from "./Footer";
-import Routes from "../routes";
+import Routes from "./routes";
 
 import "./App.less";
